requirements.txt
fastapi
uvicorn
boto3
python-dotenv
python-multipart
pydantic